Makes DrawView 'require' extension
All games will have an associated subclass of DrawView,
to handle their unique animations. Displays a message on screen
if one does not do that.

I'm fairly certain there are more graceful ways to do this but
this is what I'm going with for now. Will likely get cleaned up
later.