Sanitize current date for comparing to dates in DB
Clears the hours, minutes, seconds, and milliseconds from the current
date before calculating length of streak, as a difference in
milliseconds can cause two seemingly identical dates to be considered
different. This was not initially obvious from the toString()
represetation displayed in the debugger!

also changes oldBest++ to ++oldBest because that won't change the value
of oldBest until it's committed to the database, which is functionally
useless.