Prompts user to send a message on milestone
Compares current streak value to set of milestones, or for larger
values, checks if streak is a multiple of 50. If true, prompts user to
share their accomplishments. Intent is generic to allow using any social
media app installed on the device, or can be sent to SMS or own calendar
app.

May alter to be specific to Facebook/Twitter later.