Transitioning to standard preferences layout
Turns SettingsFragment into a PreferencesFragment, for the sake of
loading preferences.xml and inflating that to the native layout. Removes
the settings page from the main menu, because android.app.Fragment
(which is what a PreferencesFragment is) and
android.support.v4.app.Fragment (which is what flist is a list of) are
not compatible with each other.

Next step will be to add a gear icon into the main menu for accessing
the settings page.