Handles gameplay flow, just with text/screen taps
Displays the call, waits so many seconds, displays nothing, waits so
many seconds. If a tap is gotten during those two stages, a warning
is displayed and the game is canceled. If there is no tap, the response
is eventually displayed, and if a tap is received, a congratulations is
displayed and the round is considered successful. If it times out, the
game is considered failed.

Issues: Crashes when trying to update the tracker calendar, can't access
it until it's initialized by Android, which might not happen if you
haven't been all the way through the menu before launching TrainGame