Requires each game to define behaviour when ready
Makes GameActivity an abstract class and adds recognizerReady() as a
method which each Game must implement. In DeepBreathe, this is nothing
as the recognizer is not actually used and can be ignored. However for
the other games this allows them to wait until the recognizer is ready
before starting the game, potentially without voice control.

Also removes currentState and currentPair getters from TrainGame, as
they are not neccessary after refactoring.