Preliminary Start Button functionality
Keeps player from starting game until the recognizer is ready. Also
added ability to fake the recognizer being ready, for the sake of
DeepBreathe. Wanted to use a native Button for this, but it looks like
drawing that on top of a custom View like the DrawView would be a
challenge.
Use this functionality later to also show the instruction cards before
each game.