Train Game Looping & Voice Engine improvements
Train Game now loops properly, and will not pick the same word twice.
The train game has been modified to remove the concept of 'pairs',
because there aren't really that many pairs that make sense. The game is
for now a 'repeat after me' game. Some files still refer to 'pairs'
however this should not be reflected in the user interface.

The recognizer must now be passed all of the words that will be used for
the game on initialization. This allows for switching between which word
to recognize very quickly, without the error prone nature of listening
for all the words at the same time.