Adds CarGame Gameplay flow, Renames preferences
No interaction, but the car game gameplay flow is now present.

Renamed the preferences internal keys because it turns out you can't
have the same name for two different preferences, even if they're in two
different files. Game was assuming the waitTimes for TrainGame and
CarGame ought to be the same value, which they ain't