Introduces Dialog class
Changes all implementations of showDialog() to use this class. For
listeners, a new method has been added in which the listener's action
method is passed in as a Callable<Boolean>.

This means that anything that works as a listener action needs to return
a Boolean. Modified those methods to return true at the end.