Fixes Database tests
Changes the way the test manages connections so that it does not
encounter 'attempted to reopen closed file' errors. Also changed assert
values for testing incrementBest() since that doesn't actually take on
new values (in this case, 5) just increments the old value (so despite
the new value being 5, the best streak gets incremented just to 1. This
is a legacy implementation quirk)